+++ src/deluge/memory/memory_allocator_interface.cpp
@@ -1,13 +1,13 @@
 #include "general_memory_allocator.h"
 
-void* allocMaxSpeed(uint32_t requiredSize, void* thingNotToStealFrom = NULL) {
+void* allocMaxSpeed(size_t requiredSize, void* thingNotToStealFrom = NULL) {
 	return GeneralMemoryAllocator::get().alloc(requiredSize, true, false, thingNotToStealFrom);
 }
 
-void* allocLowSpeed(uint32_t requiredSize, void* thingNotToStealFrom = NULL) {
+void* allocLowSpeed(size_t requiredSize, void* thingNotToStealFrom = NULL) {
 	return GeneralMemoryAllocator::get().alloc(requiredSize, false, false, thingNotToStealFrom);
 }
 
-void* allocStealable(uint32_t requiredSize, void* thingNotToStealFrom = NULL) {
+void* allocStealable(size_t requiredSize, void* thingNotToStealFrom = NULL) {
 	return GeneralMemoryAllocator::get().alloc(requiredSize, false, true, thingNotToStealFrom);
 }
